\name{whitening-internal}



\alias{whitening-internal}


\title{Internal whitening Functions}
\description{
  Internal whitening functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
