\encoding{latin1}
\name{whitening-package}
\alias{whitening-package}
\docType{package}
\title{The whitening Package}

\description{
The "whitening" package implements the whitening methods (ZCA, PCA, Cholesky,
 ZCA-cor, and PCA-cor) discussed in Kessy, Lewin, and Strimmer (2018)
 as well as the whitening approach to canonical correlation analysis allowing negative canonical correlations described in Jendoubi and Strimmer (2018).
}

\author{Korbinian Strimmer (\url{http://strimmerlab.org/}) with Takoua Jendoubi, Agnan Kessy, and Alex Lewin.
}
\references{

Kessy, A., A. Lewin, and K. Strimmer. 2018.
Optimal whitening and decorrelation. The American Statistician.
\url{https://doi.org/10.1080/00031305.2016.1277159}

Jendoubi, T., and K. Strimmer 2018. A whitening approach to probabilistic canonical correlation analysis for omics data integration. \url{https://arxiv.org/abs/1802.03490}

Website: \url{http://strimmerlab.org/software/whitening/}
}
\keyword{multivariate}
\seealso{
\code{\link{whiteningMatrix}}.
\code{\link{whiten}},
\code{\link{cca}}, and
\code{\link{scca}}.
}
